remove duplicated logic and make room for more driveTo commands
